Toggling Franz and Steffi
Added Timers for Franz and Steffi to map them on only one button.
a: Steffi
b: Franz
The button will only change the state if there was an 1 second interval between the last change. This way we avoid it toggling constantly when the button is pushed.

There still seems to be an error as it cannot create an instance of ElapsedTime.