shooter with Encoder
enabled shooter to run with an encoder to run exactly one round. As the encoder gives 28pulses per second and an ration of 1/40 and 1/3. This means a whole rotation of the spring 5040 pulses long.
We dont want to reset the encoder to 0 once we made a whole turn because this can result in the 0 position being moved slightly with each shoot. That is why we save the last targetPosition and then add 5040 to it. This way the 0 point never gets moved.

We also want to be able to adjust the 0 point of the spring if necessary. That is why we can manually rotate the steel using gamepad2.y. We have a variable in which we set to true when we are pushing that button. As soon as we stop we can then reset the encoders and set the lastTargetPosition to 0.