wildeHilde activate only when round finished
there were some errors concerning wildeHilde. One of them was due to it setting a new targetPosition and resetting the motor when Hilde was still moving. This lead to strange behaviour. Now we also check that the currentPosition of motor is greater than the last target position. Hopefully this will fix it. Before this I had thought about an time toggle, but this solution should be better.