FrogGyro
Added custom gyro class to use. This way we should be able to read several gyros at once. For now I have kind of copy pasted the code for the range sensors and then changed the Registers to read. I will try it out later and then see if it works. Otherwise I'll need to follow this tutorial.
http://olliesworkshops.blogspot.de/2015/10/ftc-i2c-sensor-driver-architecture.html

I still need to write the functions to read.
Later I will also try out the commands to set the z_axis to zero.