MeasureDrive reset
added reset when writing a line. also added some telemetry.addData. Changed mm to cm. Set inital cm_travelled to 100.

TO DOOOOOOO
write a better runToTarget position, that slows down when nearing the end. The acceleration should be linear, I guess, so the speed will need to be x squared.