custom driveToPosition test
wanted to drive to an position and scale down over an distance. Something is not working. The scale seems to be one. I have now explicitly converted it to an double.
Later I want to create a FrogMotor, where you can set a full encoder rotation and gear ratio and things like that. This functions should be included there as well.