frogMotor calibrated
calibrated some values. When the power is lower than 0.01 it wont be moving the motor anyway. So when it reaches that state it will set the variable drivingToPosition to false and the motor will be free for usage again. Also should probably override the setPower function to ensure that there are no conflicts. Maybe also include a Timer so it will only set the variable to false when the position has been held for some time.