added FrogTableWriter
We'll probably need to write a lot of tables in the future and to be honest I just didn't want to repeat writing the same code over and over again. So I wrote a tableWriter.
It's an extension of the frogFileWriter and thus allows the same methods. For construction you need to specify the number of columns and the name of the file.
Then you can either set a specifcValue or the next value. If the columns is full please write addColumn. This will add the column and set the currentCell to 0. Also should probably reset all values to 0 here. Then call the method WriteAll to write the file. Also should probably put in some failsafes... one day....

I'll write a new class that makes sure that it displays telemetry as well, as that is a pain in the ass too.