edited FrogMotor (reset, and stopdriving)
put in a reset, that sets everything crucial to 0. (Encoder, drivingToPosition and lastTargetPosition)

also added a failsafe in initRotateRounds. It will now only be initiated when it is currently not driving to an position. Maybe I should throw an error here, to send feedback to the OpMode if that is the case.