FrogAutonomous, changed Drive
(see previous commit)
changed it to accomodate for the actionList and action stuff. initDrive and drive are now seperate entities. also included a function to return whether it is driving or not. This can be called to break the action.

ToDo
Write a simpleAutonomous, that will automatically implement the actions, when the functions are called. This will take away some options, but will later be way faster to programm new OpModes.