AutoScrimmage and driveToHeading
made a simple Auto for the scrimmage. It only drives forward and shoots two times.

driveToHeading: I don't feel like explaining what I did. But I will.
It first calculates the difference between the the expected Heading und and the current Heading. If the value is greater than 180 it subtracts 360. This way it will yield a negative value smaller than 180 and we can drive the other way around. Likewise we add 360 if the value is smaller than -180. This works. There is about an 3degrees tolerance. Will still need to try this out.