driveToWall
works alright now, but still not as great as I would like it to. It usually takes too long to react. So as with everything else, I'll need to calculate a value according to the difference of the distance measured and the desired distance. But not today!!!!

TO DO
calculate rotation with drive for TELEop
Use Gyro to drive straight.
Franz is acting weirdly