calculate rotation and driving
for this added the x value of gamepad1.rightstick (r) to b and d and substracted it from a and c. This way it should rotate if there are no x and y values, and leave it alone when there are no r values. Also this ensures that the robot can rotate and drive forward or to the side at the same time. For the scaling, it also checks if r is greater than the absolute value. If it is, it scales according to r, if not according to the absolute value of x and y.