changed omniCalculation and changeMotorDirections
now working so that if all motors are set to 1 it should drive forward. Motors diagonal of each other should this way be set to the same value. But that will require changing some code, so I'll do that later. Next thing to do will be to put the changeDirections method into the FrogOpMode, as we need that in the autonomous as well.