added adjustToHeading in TeleOpOmni4
function is similar to the driveToHeading in the Autonomous. I practically copy pasted it. Only the power is now called r_extra and added or subtracted from the drive_power values (passed as an array) accordingly, so that it is in the final value the motors are set to. Now I only need to find a good way to set the desiredHeading.