change desiredHeading in TeleOp
just changes it if r != 0 which means that the robot has been rotated intentionally. Also I should externalize all the calculation stuff. For example I could make a class DriveCalculator. Here I'd pass all the gamepad values and it would return the powers the motors have to be set to.