didn't work, tried to divide by Zero in scale
Yeah, before this it obviously returned NaN but because I only checked if all the gamepad values were not 0, I never really noticed. Also I just noticed that I probably don't need to scale by the greatest, because I scale it afterwards using the absolute value of x and y.