FrogAutonomous HeadingAdjustToFront
added functionality to adjust the heading to the front. That means that if I change the front to another side, the robot will drive so that it is pointing to the desired Heading.
Function is headingAfterFront

Also added in FrogMath degreesInCircle. If called it will ensure that the value is between 0 and 360.

So if we want to drive to a position, we call headingAfterFront, in it to adjust the desired heading.