checkSmallerLimit and driveToWall
checkSmaller takes in a value and a limit and ensures that the value is smaller than the limit (absolute values). This had to be done, because when driving to the wall we used checkSmallerOne before. This lead to it driving much too fast. Now our limit is 0.3. I will try it out in a minute.