driveToWall
sometimes the sensors would return 255. This lead to erratic behaviour. That is why we decided to set all motors to 0 should any of the sensors return 255. We then only continue driving when we get good values.
Also we would sometimes drive curves when we did not want to. To make sure that that did not happen as often, we calculate the average distance and set all motors to the same power. That way we drive straight. Only when one of the sensors returns a value smaller than the expected distance, do we calculate seperate values for the left and right side.