added FrogAccelerometer
up until now, we can read the accelerometer values and also set the Accelerometer scale. We should probably put in something into FrogMath to convert ints to fsb/lsb format.
For the fsb/lsb format see the Modern robotics documentation.

Next thing to do will be to write a test.