added stop for franz and steffi
as these two motors would reverse after about 40s we decided to solve this problem using software. For this we use a elapsedTime that is always reset when the toggle returns false (when the motor isn't turning) This means, that is measures the time since the motor was activated. When the time is greater than 40 seconds it will set the power to 0. We hope that this will be enough to prevent the motor from reversing.