fixed wildeHildeMotor auto rotation
probably caused erratic behaviour because states from last time where saved. To solve this we reset everythin in the init. Also heightended the tolerance for stopping. This was probably changed, because we use a different maximum speed or something.

I cant remember setting a maximum speed, so maybe that's not true.

Still need to try this out and run a lot of tests.