hildeEncMeasure
added class to get the current position of the encoder measurements of hilde. This is to set the maximum speed for hilde to enable better driving. This is especially important for the auto rotate round functionality. It writes to the file hilde_enc.csv

Buttons:
a starts rotation
b stops rotation
x saves the file