maxSpeed of hilde
Measurements yielded an average value of 2848 ticks per second. So we set the max speed to that in the FrogOpMode.
After testing this out, we found out that it had just slightly slowered the motor. So we increased the value to 3000.
It works pretty well and the motor rotates much better now.
This is because the scaling is more linear now than it was before.