wildeHildeMotor rotate backwards, some conditions
it will now rotate backward when both y and x are pressed.
To make sure that it does not try to initiate rounds when y is pressed as well (eg wanting to drive backwards) we added a condition to check this when initiating the round.
Also we added an condition in front of the driveToPosition. It will now only do this when y is not pressed. This is to make sure that driving to a position and moving something manually do not conflict.

Hilde works pretty neat now.