Omnidiretional Drive Encoder Data
We want to be able to drive into all directions using our encoders.
That is why we need a lot of data, as we do not know how the motors interact when not driving forward.
We have thought about changing the value of one diagonal motor pair, while leaving the others at 1.
Then we would measure the encoder values and compare that to the ratio the motors were set to.
Obviously we want the x and y distance the robot travelled. This is easier to measure than the angle and the total distance.

Here's what the table will probably look like.
Index, ratio1, ratio2, A, B, C, D, X_distance, Y_distance

This is implemented in the MeasureOmniDrive class.