Bugfixes in MeasureOmniDrive
Now decreasing the motorPowerBD with each measurement taken.
Tested it out, seems to be working alright.
Tomorrow we will try it out on the playing field and take real measurements to work with.

What we will do with the measurements
This will enable us to drive diagonally as well giving us more freedom and saving us a lot of time.

The next thing to do will then be to get this to work well when driving between two points.
After we have achieved this, we should again take a look at the route calculation. As we already have most of the theory and driving between two points is taken care of.
Another thing to do is the javadocs.