Changed max speed of drive motors
Since our motors aren't driving straight, we decided to take a new look at the speed measurements.
Before we had calculated the average speed of all motors. This time we calculated them for each motor individually.
This returned that the average speed varied from motor to motor.
The slowest motor (aOmni) turns at 1649 pulses per second. That is slower than the 1677 we set the maxSpeed to.
That is why we have set the maximum speed of all motors to 1600 pulses per second.