Driving straight, scaling the motors differently
We aren't able to drive straight. One approach to solve that problem is to scale the powers according to the motors.
This is to compensate for the hardware. To do that we took some measurements, setting all the powers to 1.
We then evaluated the encoder data and calculated the diversion from the average for each of them.
We then scaled the maximum speed of each motor according to the calculated values.#