FrogVector for autonomous
since we will be using vectors more and more, we decided to put all the functionality concerning them into a single class.
This makes it a lot easier to deal with them.
It converts x y values to polar representation and back.
Also there is functionality to add a vector.