Bugfixes on calculatedHeading
the calculatedHeading is always set when the heading is changed. It is important for calculcating how the robot needs to drive.
We forgot to set the calculatedHeading in the beginning.This lead to weird behaviour.
Later we want to enable the robot to use the current heading and calculate routes "on the fly".
For now it seems to work, but there is still much room for improvement. It still needs fine tuning.
There are probably some bugs hiding there as well.