A range sensor workaround
Our range sensors often return 255 as values. When this happens, we just stopped the robot.
This lead to our robot wasting time in the autonomous. We wanted to change that.

One approach was to actually solve the problem. But we did not find the source. Trying to solve intereference problems was hard, as can be seen in older commits.
Reading in the teleOp showed that the problem persisted.

We have decided to use a workaround for now.
If the sensors return 255 we just return the older values. Only if the problem persists for longer than a set time interval do we actually stop the robot.

To implement this, we decided to put everything into a new class named the FrogBeaconRangeSensors class.
This is to avoid confusing behaviour when using the FrogRange class. That class is only meant to return the read value, without processing.
Int the FrogBeaconRangeSensors class, we have methods to read the left and right values.
Then we use the method described above. If the set time interval has passed and the sensor is still returning 255, these functions return 255.
Otherwise the older value is returned.

Everything in the FrogAutonomous thus remains mostly the same. Only the values are gotten through the FrogBeaconRangeSensors class.