Game clock
since the game clock will be used by all op modes, we have implemented it in the FrogOpMode.
Here we rotate the pointer for a set amount of time and then stop it until 5 seconds have passed. Then we continue rotating it.
It's not the most accurate method, but we can't have it better since setting the power to a lower level doesn't rotate the servo at all.
It should be enough to provide an overview of the time. Nevertheless a coach should alwaysbe watching the real game clock.